OpenAMP: Add openAMP master support
The openAMP is a master-slave mechanism. TFM supports this feature by
including external Libraries. TFM act as master in secure core,
applications act as slaves in non-secure cores. And the slave better to
keep the same version with master for compatbility.

This patch will allow TFM import openAMP from Zephyr's repo, as the
upcoming examples are all cooperate with Zephyr in non-secure core.

Change-Id: Ib4447e54824604eca08c5610c97767e07ae35b2e
Signed-off-by: Karl Zhang <karl.zhang@linaro.org>