eRPC: Compile eRPC as external library
eRPC (Embedded RPC) is an open source Remote Procedure Call (RPC) system
for multichip embedded systems and heterogeneous multicore SoCs.

The repo using here is based on:
    https://github.com/EmbeddedRPC/erpc
    commit: a5bb96821c111164defa6c8c932e189a8bcf86c2

The eRPC in TF-M act as server, other cores in the system will be
clients. The modifications in eRPC was not ready for PR yet, checkout
eRPC from the github below before the changes merged into official repo:
    https://github.com/karl-zh/erpc.git -b build_with_tfm_for_cm33

Change-Id: I5a81d41c55fae9355545ad50180dfd337b95037d